+++ component/prometheus/remotewrite/remote_write.go
@@ -8,6 +8,8 @@ import (
 	"sync"
 	"time"
 
+	"go.uber.org/atomic"
+
 	"github.com/prometheus/prometheus/model/exemplar"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/model/metadata"
@@ -50,6 +52,7 @@ type Component struct {
 	walStore    *wal.Storage
 	remoteStore *remote.Storage
 	storage     storage.Storage
+	exited      atomic.Bool
 
 	mut sync.RWMutex
 	cfg Arguments
@@ -86,6 +89,10 @@ func NewComponent(o component.Options, c Arguments) (*Component, error) {
 		// ID" to ensure Flow compatibility.
 
 		prometheus.WithAppendHook(func(globalRef storage.SeriesRef, l labels.Labels, t int64, v float64, next storage.Appender) (storage.SeriesRef, error) {
+			if res.exited.Load() {
+				return 0, fmt.Errorf("%s has exited", o.ID)
+			}
+
 			localID := prometheus.GlobalRefMapping.GetLocalRefID(res.opts.ID, uint64(globalRef))
 			newRef, nextErr := next.Append(storage.SeriesRef(localID), l, t, v)
 			if localID == 0 {
@@ -94,6 +101,10 @@ func NewComponent(o component.Options, c Arguments) (*Component, error) {
 			return globalRef, nextErr
 		}),
 		prometheus.WithMetadataHook(func(globalRef storage.SeriesRef, l labels.Labels, m metadata.Metadata, next storage.Appender) (storage.SeriesRef, error) {
+			if res.exited.Load() {
+				return 0, fmt.Errorf("%s has exited", o.ID)
+			}
+
 			localID := prometheus.GlobalRefMapping.GetLocalRefID(res.opts.ID, uint64(globalRef))
 			newRef, nextErr := next.UpdateMetadata(storage.SeriesRef(localID), l, m)
 			if localID == 0 {
@@ -102,6 +113,10 @@ func NewComponent(o component.Options, c Arguments) (*Component, error) {
 			return globalRef, nextErr
 		}),
 		prometheus.WithExemplarHook(func(globalRef storage.SeriesRef, l labels.Labels, e exemplar.Exemplar, next storage.Appender) (storage.SeriesRef, error) {
+			if res.exited.Load() {
+				return 0, fmt.Errorf("%s has exited", o.ID)
+			}
+
 			localID := prometheus.GlobalRefMapping.GetLocalRefID(res.opts.ID, uint64(globalRef))
 			newRef, nextErr := next.AppendExemplar(storage.SeriesRef(localID), l, e)
 			if localID == 0 {
@@ -128,6 +143,8 @@ var _ component.Component = (*Component)(nil)
 // Run implements Component.
 func (c *Component) Run(ctx context.Context) error {
 	defer func() {
+		c.exited.Store(true)
+
 		level.Debug(c.log).Log("msg", "closing storage")
 		err := c.storage.Close()
 		level.Debug(c.log).Log("msg", "storage closed")