+++ CHANGELOG.md
@@ -56,6 +56,8 @@ Main (unreleased)
 
 - Flow UI: Fix the issue with long string going out of bound in the component detail page. (@xiyu95)
 
+- Flow: `prometheus.relabel` and `prometheus.remote_write` will now error if they have exited. (@ptodev)
+
 ### Other changes
 
 - Use Go 1.19.4 for builds. (@erikbaranowski)